Compare ints to ints when looking up pages
`input.id` can sometimes be a string (from API?) and throws a comparison error here. TODO: make sure `input.aspect_ratio` can survive being a string a few lines up; should be okay b'c it's assignment, not comparison, but check.

+++ app/actions/calculate_aspect_ratios.rb
@@ -30,7 +30,7 @@ class CalculateAspectRatios < CloudCrowd::Action
     UPDATE pages 
       SET aspect_ratio = input.aspect_ratio
       FROM (SELECT unnest(ARRAY[?]) as id, unnest(ARRAY[?]) as aspect_ratio) AS input
-      WHERE pages.id = input.id
+      WHERE pages.id = input.id::int
     QUERY
     query = Page.send(:replace_bind_variables, query_template, [ids, @page_aspect_ratios])
     Page.connection.execute(query)
